Sharing purpose of visit logic with channel island question
Here I have moved the the blocks and logic within the purpose of visit multiple choice question to the travel_response_next_route method (an instance method on CheckUkVisaFlow).

The reason for this is mainly because, as part of the new requirement for the i
Common travel areas (CTA).

A three new options has been introduced (replacing the previous yes/no) to
identify the two main CTA territories and rules that apply.

For users travelling to the Channel island or isle of man they will be presente
with purpose of visit question without the transit option and the logic/routes that follow are exaclty or a replay of the purpose of visit.

The logic here remains the same with purpose of visit (i.e Question 2).

Consequently, fixed failing tests